+++ src/_tests_/createAppContainer.roblox.spec.lua
@@ -4,8 +4,14 @@ return function()
 	local Roact = require(Packages.Roact)
 	local jestExpect = require(Packages.Dev.JestRoblox).Globals.expect
 	local NavigationActions = require(RoactNavigationModule.NavigationActions)
-	local createAppContainer = require(RoactNavigationModule.createAppContainer)
+	local createAppContainerExports = require(RoactNavigationModule.createAppContainer)
+	local createAppContainer = createAppContainerExports.createAppContainer
+	local _TESTING_ONLY_reset_container_count = createAppContainerExports._TESTING_ONLY_reset_container_count
 	local createRobloxSwitchNavigator = require(RoactNavigationModule.navigators.createRobloxSwitchNavigator)
+
+	beforeEach(function()
+		_TESTING_ONLY_reset_container_count()
+	end)
 
 	it("should be a function", function()
 		jestExpect(createAppContainer).toEqual(jestExpect.any("function"))
@@ -127,5 +133,140 @@ return function()
 
 		Roact.unmount(instance)
 	end)
-end
 
+	describe("with deep linking", function()
+		local function getLinkingProtocolMock(initialURL)
+			local linkingProtocolMock = {}
+
+			function linkingProtocolMock:listenForLuaURLs(callback, sticky)
+				self.callback = callback
+				self.sticky = sticky
+			end
+
+			function linkingProtocolMock:getLastLuaURL()
+				return initialURL
+			end
+
+			function linkingProtocolMock:stopListeningForLuaURLs()
+				self.callback = nil
+			end
+
+			return linkingProtocolMock
+		end
+
+		local function findFirstDescendantOfClass(parent, className)
+			for _, descendant in ipairs(parent:GetDescendants()) do
+				if descendant:IsA(className) then
+					return descendant
+				end
+			end
+			return nil
+		end
+
+		-- use a class that we will never find within the instance hierarchy created
+		-- by Roact Navigation. That way if the swith navigation implementation changes,
+		-- we won't have to modify the test
+		local UNIQUE_CLASS_NAME = "HopperBin"
+
+		it("connects and disconnects from `listenForLuaURLs`", function()
+			local function Screen(_props)
+				return nil
+			end
+			local testNavigator = createRobloxSwitchNavigator({
+				{ Foo = { screen = Screen, path = "foo" } },
+				{ Bar = { screen = Screen, path = "bar" } },
+			})
+
+			local protocolMock = getLinkingProtocolMock("foo")
+			local app = createAppContainer(testNavigator, protocolMock)
+
+			local element = Roact.createElement(app)
+			local tree = Roact.mount(element)
+
+			jestExpect(protocolMock.callback).never.toEqual(nil)
+			jestExpect(protocolMock.sticky).toEqual(false)
+
+			Roact.unmount(tree)
+
+			jestExpect(protocolMock.callback).toEqual(nil)
+		end)
+
+		it("uses the last URL to set the initial navigation state", function()
+			local fooElementClass = "TextLabel"
+			local barElementClass = "Frame"
+			local function FooScreen(_props)
+				return Roact.createElement(UNIQUE_CLASS_NAME, {}, {
+					Foo = Roact.createElement(fooElementClass),
+				})
+			end
+			local function BarScreen(_props)
+				return Roact.createElement(UNIQUE_CLASS_NAME, {}, {
+					Bar = Roact.createElement(barElementClass)
+				})
+			end
+
+			for url, expectedClass in pairs({
+				foo = {fooElementClass, barElementClass},
+				bar = {barElementClass, fooElementClass},
+			}) do
+				local testNavigator = createRobloxSwitchNavigator({
+					{ Foo = { screen = FooScreen, path = "foo" } },
+					{ Bar = { screen = BarScreen, path = "bar" } },
+				})
+
+				local protocolMock = getLinkingProtocolMock(url)
+				local app = createAppContainer(testNavigator, protocolMock)
+
+				local element = Roact.createElement(app)
+				local parent = Instance.new("Folder")
+				local tree = Roact.mount(element, parent)
+
+				local screen = findFirstDescendantOfClass(parent, UNIQUE_CLASS_NAME)
+				jestExpect(screen).toBeDefined()
+				jestExpect(screen:FindFirstChildOfClass(expectedClass[1])).toBeDefined()
+				jestExpect(screen:FindFirstChildOfClass(expectedClass[2])).toBeUndefined()
+
+				Roact.unmount(tree)
+			end
+		end)
+
+		it("can get the params from the initial URL", function()
+			local fooElementClass = "TextLabel"
+			local barElementClass = "Frame"
+			local function FooScreen(_props)
+				return Roact.createElement(UNIQUE_CLASS_NAME, {}, {
+					Foo = Roact.createElement(fooElementClass),
+				})
+			end
+			local function BarScreen(props)
+				local navigation = props.navigation
+				local name = navigation.getParam("name")
+				return Roact.createElement(UNIQUE_CLASS_NAME, {}, {
+					[name] = Roact.createElement(barElementClass)
+				})
+			end
+
+			local testNavigator = createRobloxSwitchNavigator({
+				{ Foo = { screen = FooScreen, path = "foo" } },
+				{ Bar = { screen = BarScreen, path = "bar/:name" } },
+			})
+
+			local expectName = "orange"
+			local protocolMock = getLinkingProtocolMock("bar/" .. expectName)
+			local app = createAppContainer(testNavigator, protocolMock)
+
+			local element = Roact.createElement(app)
+			local parent = Instance.new("Folder")
+			local tree = Roact.mount(element, parent)
+
+			local screen = findFirstDescendantOfClass(parent, UNIQUE_CLASS_NAME)
+			jestExpect(screen).toBeDefined()
+			local barInstance = screen:FindFirstChildOfClass(barElementClass)
+			jestExpect(barInstance).toBeDefined()
+			jestExpect(barInstance.Name).toEqual(expectName)
+			jestExpect(screen:FindFirstChildOfClass(fooElementClass)).toBeUndefined()
+
+			Roact.unmount(tree)
+		end)
+	end)
+end